move header dropdown in specific component
