+++ src/components/Navigation/Sections/HeaderSection.js
@@ -5,17 +5,13 @@ import PropTypes from 'prop-types'
 import { Row, Button, Col } from 'react-bootstrap'
 import isEmpty from 'lodash/isEmpty'
 import get from 'lodash/get'
-import { Tag, Dropdown, Menu, Icon } from 'antd'
+import { Tag } from 'antd'
 import { withResize } from '../../../utils/WindowProvider'
 import Curation from '../../../utils/curation'
 import ScoreRenderer from '../Ui/ScoreRenderer'
 import DefinitionTitle from '../Ui/DefinitionTitle'
 import DefinitionRevision from '../Ui/DefinitionRevision'
-import CopyUrlButton from '../../CopyUrlButton'
-import { ROUTE_DEFINITIONS } from '../../../utils/routingConstants'
-import EntitySpec from '../../../utils/entitySpec'
-import Definition from '../../../utils/definition'
-import ButtonWithTooltip from '../Ui/ButtonWithTooltip'
+import ComponentDetailsButtons from '../Ui/ComponentDetailsButtons'
 
 class HeaderSection extends Component {
   static propTypes = {
@@ -27,22 +23,6 @@ class HeaderSection extends Component {
     handleClose: PropTypes.func,
     handleSave: PropTypes.func,
     handleRevert: PropTypes.func
-  }
-
-  constructor(props) {
-    super(props)
-    this.copyUrlButton = React.createRef()
-  }
-
-  isSourceComponent(component) {
-    return ['github', 'sourcearchive'].includes(component.provider)
-  }
-
-  openSourceForComponent = definition => {
-    const sourceLocation = get(definition, 'described.sourceLocation')
-    const sourceEntity = sourceLocation && EntitySpec.fromObject(sourceLocation)
-    const path = `${ROUTE_DEFINITIONS}/${EntitySpec.fromObject(sourceEntity).toPath()}`
-    window.open(`${window.location.origin}${path}`)
   }
 
   render() {
@@ -62,37 +42,6 @@ class HeaderSection extends Component {
     const isCurated = Curation.isCurated(curations.item)
     const hasPendingCurations = Curation.isPending(curations.item)
 
-    const menu = (
-      <Menu>
-        <Menu.Item key="1" onClick={() => this.copyUrlButton.current.onCopy()}>
-          Copy Component Url{' '}
-          <CopyUrlButton
-            ref={this.copyUrlButton}
-            className="list-fa-button"
-            path={`${ROUTE_DEFINITIONS}/${EntitySpec.fromObject(get(item, 'coordinates')).toPath()}`}
-          />
-        </Menu.Item>
-        {!isSourceComponent && !isSourceEmpty && (
-          <Menu.Item key="2" onClick={() => this.openSourceForComponent(item)}>
-            Open Source Definition{' '}
-            <ButtonWithTooltip tip="Open the definition for source that matches this package" placement="bottom">
-              <Button className="list-fa-button">
-                <i className="fas fa-code" />
-              </Button>
-            </ButtonWithTooltip>
-          </Menu.Item>
-        )}
-        <Menu.Item key="3">
-          List other version of this component{' '}
-          <Button className="list-fa-button">
-            <i class="fas fa-list" />
-          </Button>
-        </Menu.Item>
-      </Menu>
-    )
-
-    const isSourceComponent = this.isSourceComponent(item.coordinates)
-    const isSourceEmpty = Definition.isSourceEmpty(item)
     return (
       <Row className="row-detail-header">
         <Col md={8}>
@@ -120,11 +69,7 @@ class HeaderSection extends Component {
                 )}
               </div>
               <div>
-                <Dropdown overlay={menu} trigger="hover">
-                  <Button bsStyle="info">
-                    <i class="fas fa-ellipsis-v" />
-                  </Button>
-                </Dropdown>
+                <ComponentDetailsButtons item={item} />
               </div>
             </div>
             <p>
